requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
click-default-group==1.2.4
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
distro==1.9.0
faiss-cpu==1.11.0
fastapi==0.115.12
fonttools==4.58.4
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
langchain==0.3.25
langchain-core==0.3.65
langchain-openai==0.3.23
langchain-text-splitters==0.3.8
langgraph==0.4.8
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.2.2
langgraph-sdk==0.1.70
langsmith==0.3.45
MarkupSafe==3.0.2
matplotlib==3.10.3
narwhals==1.42.1
numpy==2.3.0
openai==1.86.0
orjson==3.10.18
ormsgpack==1.10.0
packaging==24.2
pandas==2.3.0
pillow==11.2.1
plotly==6.1.2
pluggy==1.6.0
protobuf==6.31.1
psycopg2-binary==2.9.10
pyarrow==20.0.0
pydantic==2.11.6
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.3
pytest==8.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
requests-toolbelt==1.0.0
rpds-py==0.25.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlite-fts4==1.0.3
sqlite-utils==3.38
starlette==0.46.2
streamlit==1.45.1
tabulate==0.9.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.3
watchdog==6.0.0
xxhash==3.5.0
zstandard==0.23.0

kaleido==0.1.*